requirements.txt
asgiref==3.7.2
Django==5.0.2
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.4.4
